axios error handling #validation

+++ frontend/src/App.js
@@ -2,25 +2,52 @@ import React, { useEffect, useState } from "react";
 import axios from "axios";
 import "./App.css";
 import { Link } from "react-router-dom";
-
+import Loading from "./components/Loading";
 const App = () => {
   const [tasks, setTasks] = useState([]);
+  const [LoadingTrue, setLoadingTrue] = useState(false);
 
   // Fetch tasks from the backend
   useEffect(() => {
+    setLoadingTrue(true);
     const fetchTasks = async () => {
-      const response = await axios.get("http://localhost:5000/api/tasks");
-      setTasks(response.data);
+      try {
+        const response = await axios.get("http://localhost:5000/api/tasks");
+        setTasks(response.data);
+        setLoadingTrue(false);
+      } catch (error) {
+        alert("Error fetching tasks. Please try again.");
+        console.error("Fetch error: ", error);
+        setLoadingTrue(false);
+      }
     };
     fetchTasks();
   }, []);
 
   // Delete task
   const deleteTask = async (id) => {
-    await axios.delete(`http://localhost:5000/api/tasks/${id}`);
-    setTasks(tasks.filter((task) => task.id !== id));
-  };
+    const confirmDelete = window.confirm(
+      "Are you sure you want to delete this task?"
+    );
 
+    if (confirmDelete) {
+      setLoadingTrue(true);
+
+      try {
+        await axios.delete(`http://localhost:5000/api/tasks/${id}`);
+        setTasks(tasks.filter((task) => task.id !== id));
+        alert("Task deleted successfully.");
+        setLoadingTrue(false);
+      } catch (error) {
+        alert("Error deleting task. Please try again.");
+        console.error("Delete error: ", error);
+        setLoadingTrue(false);
+      }
+    }
+  };
+  if (LoadingTrue) {
+    return <Loading />;
+  }
   return (
     <div className="p-6 bg-slate-100 min-h-[100vh]">
       <section className="flex items-baseline gap-10">
